// src/app/models/guestview/Corporate.ts
import {CoHostInCharge} from './PropertyForGuest';

export class Corporate {
  id: string;
  name: string;
  corporateTag: string;

  fullAddress: string;
  address: string;
  zipCode: string;
  city: string;
  country: string;
  email: string;
  phone: string;
  facebookUrl: string;
  instagramUrl: string;
  twitterUrl: string;
  linkedInUrl: string;

  primaryColor: string;
  secondaryColor: string;
  forgetMeNotColor: string;
  greenVogueColor: string;
  catskillColor: string;
  logoUrl: string;
  onlyWhiteLabel: boolean;
  coHostInCharge: CoHostInCharge;
  showMeAsHost: boolean;

  yaagoFooterBackgroundColor: string;
  yaagoFooterTextColor: string;

  bookletMasterKey: string;
}